requirements.txt
netmiko
colorama
termcolor
dnspython
